requirements.txt
python-dotenv
requests
whisperx
sounddevice
soundfile
